requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
feedparser==6.0.11
python-dateutil==2.8.2
requests==2.31.0
django-filter==23.5
dj-rest-auth==5.0.2
django-allauth==0.61.1
gunicorn==21.2.0
whitenoise==6.6.0 